[CYTHON] Enable ffi function creation from MLIR Packed function
The MLIR execution engine generates functions that correspond to the packed signature.
As of now, it is hard to access the raw extern C function pointer of SafeCall
directly when we declare the signature in LLVM dialect.

Note that in theory, the MLIR execution engine should be able to support
some form of "extern C" feature that directly exposes the function pointers
of C-compatible functions with an attribute tag. So we keep this feature
in the Python helper layer for now in case the MLIR execution engine supports it in the future.